+++ internal/task/noop.go
@@ -1,4 +1,4 @@
-package upload
+package task
 
 type NoOpJob struct {}