+++ internal/upload/types.go
@@ -1,17 +1,5 @@
 package upload
 
-import (
-	"context"
-
-	"github.com/gphotosuploader/googlemirror/api/photoslibrary/v1"
-)
-
-// gPhotosService represents a Google Photos Service.
-type gPhotosService interface {
-	GetOrCreateAlbumByName(name string) (*photoslibrary.Album, error)
-	AddMediaItem(ctx context.Context, path string, album string) (*photoslibrary.MediaItem, error)
-}
-
 // FileTracker represents a service to track already uploaded files.
 type FileTracker interface {
 	CacheAsAlreadyUploaded(filePath string) error