+++ internal/upload/walker.go
@@ -3,31 +3,26 @@ package upload
 import (
 	"os"
 	"path/filepath"
+	"strings"
 
 	"github.com/gphotosuploader/gphotos-uploader-cli/internal/log"
-	"github.com/gphotosuploader/gphotos-uploader-cli/internal/utils/filesystem"
 )
-
-type UploadItem struct {
-	Path      string
-	AlbumName string
-}
 
 // ScanFolder return the list of Items{} to be uploaded. It scans the folder and skip
 // non allowed files (includePatterns & excludePattens).
-func (job *UploadFolderJob) ScanFolder(logger log.Logger) ([]UploadItem, error) {
-	var result []UploadItem
+func (job *UploadFolderJob) ScanFolder(logger log.Logger) ([]FileItem, error) {
+	var result []FileItem
 	err := filepath.Walk(job.SourceFolder, job.getItemToUploadFn(&result, logger))
 	return result, err
 }
 
-func (job *UploadFolderJob) getItemToUploadFn(reqs *[]UploadItem, logger log.Logger) filepath.WalkFunc {
+func (job *UploadFolderJob) getItemToUploadFn(reqs *[]FileItem, logger log.Logger) filepath.WalkFunc {
 	return func(fp string, fi os.FileInfo, errP error) error {
 		if fi == nil {
 			return nil
 		}
 
-		relativePath := filesystem.RelativePath(job.SourceFolder, fp)
+		relativePath := RelativePath(job.SourceFolder, fp)
 
 		// If a directory is excluded, skip it!
 		if fi.IsDir() {
@@ -59,10 +54,24 @@ func (job *UploadFolderJob) getItemToUploadFn(reqs *[]UploadItem, logger log.Log
 		logger.Infof("File '%s' will be uploaded to album '%s'.", fp, job.albumName(relativePath))
 
 		// set file upload Options depending on folder upload Options
-		*reqs = append(*reqs, UploadItem{
+		*reqs = append(*reqs, FileItem{
 			Path:      fp,
 			AlbumName: job.albumName(relativePath),
 		})
 		return nil
 	}
+}
+
+// RelativePath returns a path relative to the base.
+// If a relative path could not be calculated or it contains ' ../`,
+// returns the original path.
+func RelativePath(base string, path string) string {
+	rp, err := filepath.Rel(base, path)
+	if err != nil {
+		return path
+	}
+	if strings.HasPrefix(rp, "../") {
+		return path
+	}
+	return rp
 }